requirements.txt
# === Core PyTorch ===
torch            # Pin version if using CUDA-specific builds
torchvision
torchaudio

# === Transformers & CLIP ===
transformers>=4.30.0
sentencepiece>=0.1.99
ftfy>=6.1.1
regex
tqdm

# === Image & Video Processing ===
ultralytics
Pillow>=9.0.0
opencv-python>=4.5.5
imageio
imageio-ffmpeg

# === Data & Evaluation Utilities ===
numpy
scikit-learn
pandas
openpyxl
seaborn
tensorboard

# === Visualization & Notebook ===
matplotlib
ipykernel

gdown